requirements.txt
Flask==3.1.1
ftfy==6.3.1
numpy==2.3.0
Pillow==11.2.1
regex==2024.11.6
setuptools==57.4.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.27.3
